requirements.txt
PyMuPDF==1.23.21
faiss-cpu==1.7.4
numpy==1.26.4
sentence-transformers==2.6.1
openai==1.30.1